+++ dataset/freeze/format/fcsv.py
@@ -8,15 +8,7 @@ from dataset.freeze.format.common import Serializer
 
 
 def value_to_str(value):
-    if isinstance(value, datetime):
-        #
-        # FIXME: this check does not work for values returned from a db query!
-        # As a workaround, we make sure, the isoformat call returns the regular
-        # str representation.
-        #
-        sep = ' ' if PY3 else str(' ')
-        return text_type(value.isoformat(sep=sep))
-    if isinstance(value, date):
+    if isinstance(value, (date, datetime)):
         return text_type(value.isoformat())
     if not PY3 and hasattr(value, 'encode'):
         return value.encode('utf-8')
@@ -33,7 +25,6 @@ class CSVSerializer(Serializer):
     def write(self, path, result):
         keys = list(result.keys())
         if path not in self.handles:
-
             # handle fileobj that has been passed in:
             if path is not None:
                 if PY3:  # pragma: no cover
@@ -47,7 +38,7 @@ class CSVSerializer(Serializer):
             if PY3:  # pragma: no cover
                 writer.writerow(keys)
             else:
-                writer.writerow([k.encode('utf-8') for k in keys])
+                writer.writerow([value_to_str(k) for k in keys])
             self.handles[path] = (writer, fh)
         writer, fh = self.handles[path]
         values = [value_to_str(result.get(k)) for k in keys]
@@ -55,4 +46,5 @@ class CSVSerializer(Serializer):
 
     def close(self):
         for writer, fh in self.handles.values():
-            fh.close()
+            if fh != self.fileobj:
+                fh.close()