+++ dataset/freeze/format/fjson.py
@@ -25,9 +25,9 @@ class JSONSerializer(Serializer):
     def wrap(self, result):
         if self.mode == 'item':
             result = result[0]
-        if self.wrap:
+        if self.export.get_bool('wrap', True):
             result = {
-                'count': self.query.count,
+                'count': len(result),
                 'results': result
             }
             meta = self.export.get('meta', {})
@@ -50,9 +50,11 @@ class JSONSerializer(Serializer):
             data = json.dumps(result,
                               cls=JSONEncoder,
                               indent=self.export.get_int('indent'))
-            if self.export.get('callback'):
-                data = "%s && %s(%s);" % (self.export.get('callback'),
-                                          self.export.get('callback'),
-                                          data)
+
+            callback = self.export.get('callback')
+            if callback:
+                data = "%s && %s(%s);" % (callback, callback, data)
+
             fh.write(data)
-            fh.close()
+            if self.fileobj is None:
+                fh.close()