+++ dataset/persistence/util.py
@@ -1,12 +1,18 @@
-from datetime import datetime, timedelta
-from inspect import isgenerator
+from datetime import datetime
+try:
+    from urlparse import urlparse
+except ImportError:
+    from urllib.parse import urlparse
 
 try:
     from collections import OrderedDict
 except ImportError:  # pragma: no cover
     from ordereddict import OrderedDict
 
-from sqlalchemy import Integer, UnicodeText, Float, DateTime, Boolean, types, Table, event
+from sqlalchemy import Integer, UnicodeText, Float, DateTime, Boolean
+from six import string_types
+
+row_type = OrderedDict
 
 
 def guess_type(sample):
@@ -21,44 +27,54 @@ def guess_type(sample):
     return UnicodeText
 
 
-def convert_row(row):
+def convert_row(row_type, row):
     if row is None:
         return None
-    return OrderedDict(row.items())
+    return row_type(row.items())
+
+
+def normalize_column_name(name):
+    if not isinstance(name, string_types):
+        raise ValueError('%r is not a valid column name.' % name)
+    name = name.lower().strip()
+    if not len(name) or '.' in name or '-' in name:
+        raise ValueError('%r is not a valid column name.' % name)
+    return name
 
 
 class ResultIter(object):
     """ SQLAlchemy ResultProxies are not iterable to get a
     list of dictionaries. This is to wrap them. """
 
-    def __init__(self, result_proxies):
-        if not isgenerator(result_proxies):
-            result_proxies = iter((result_proxies, ))
-        self.result_proxies = result_proxies
-        self.count = 0
-        self.rp = None
+    def __init__(self, result_proxy, row_type=row_type, step=None):
+        self.result_proxy = result_proxy
+        self.row_type = row_type
+        self.step = step
+        self.keys = list(result_proxy.keys())
+        self._iter = None
 
-    def _next_rp(self):
-        try:
-            self.rp = next(self.result_proxies)
-            self.count += self.rp.rowcount
-            self.keys = list(self.rp.keys())
+    def _next_chunk(self):
+        if self.result_proxy.closed:
+            return False
+        if not self.step:
+            chunk = self.result_proxy.fetchall()
+        else:
+            chunk = self.result_proxy.fetchmany(self.step)
+        if chunk:
+            self._iter = iter(chunk)
             return True
-        except StopIteration:
+        else:
             return False
 
     def __next__(self):
-        if self.rp is None:
-            if not self._next_rp():
+        if self._iter is None:
+            if not self._next_chunk():
                 raise StopIteration
-        row = self.rp.fetchone()
-        if row is None:
-            if self._next_rp():
-                return next(self)
-            else:
-                # stop here
-                raise StopIteration
-        return convert_row(row)
+        try:
+            return convert_row(self.row_type, next(self._iter))
+        except StopIteration:
+            self._iter = None
+            return self.__next__()
 
     next = __next__
 
@@ -66,32 +82,10 @@ class ResultIter(object):
         return self
 
 
-def sqlite_datetime_fix():
-    class SQLiteDateTimeType(types.TypeDecorator):
-        impl = types.Integer
-        epoch = datetime(1970, 1, 1, 0, 0, 0)
-
-        def process_bind_param(self, value, dialect):
-            if value is None:
-                return None
-            if isinstance(value, datetime):
-                return value
-            return (value / 1000 - self.epoch).total_seconds()
-
-        def process_result_value(self, value, dialect):
-            if value is None:
-                return None
-            if isinstance(value, int):
-                return self.epoch + timedelta(seconds=value / 1000)
-            return value
-
-    def is_sqlite(inspector):
-        return inspector.engine.dialect.name == "sqlite"
-
-    def is_datetime(column_info):
-        return isinstance(column_info['type'], types.DateTime)
-
-    @event.listens_for(Table, "column_reflect")
-    def setup_epoch(inspector, table, column_info):
-        if is_sqlite(inspector) and is_datetime(column_info):
-            column_info['type'] = SQLiteDateTimeType()
+def safe_url(url):
+    """Remove password from printed connection URLs."""
+    parsed = urlparse(url)
+    if parsed.password is not None:
+        pwd = ':%s@' % parsed.password
+        url = url.replace(pwd, ':*****@')
+    return url