+++ test/test_freeze_app.py
@@ -1,8 +1,9 @@
+# coding: utf-8
 """
 Test CLI following the recipe at http://dustinrcollins.com/testing-python-command-line-apps
 """
 import os
-from unittest import TestCase
+import unittest
 from tempfile import mkdtemp
 from shutil import rmtree
 from copy import copy
@@ -16,7 +17,7 @@ from dataset.freeze.app import create_parser, freeze_with_config, freeze_export
 from .sample_data import TEST_DATA
 
 
-class FreezeTestCase(TestCase):
+class FreezeAppTestCase(unittest.TestCase):
     """
     Base TestCase class, sets up a CLI parser
     """
@@ -38,15 +39,7 @@ class FreezeTestCase(TestCase):
     def tearDown(self):
         rmtree(self.d, ignore_errors=True)
 
-    def test_with_empty_args(self):
-        """
-        User passes no args, should fail with SystemExit
-        """
-        self.assertRaises(SystemExit, self.parser.parse_args, [])
-
     def test_with_config(self):
-        """
-        """
         cfg = Configuration(os.path.join(os.path.dirname(__file__), 'Freezefile.yaml'))
         cfg.data['common']['database'] = self.db
         cfg.data['common']['prefix'] = self.d
@@ -60,3 +53,15 @@ class FreezeTestCase(TestCase):
             {'skip': True}]
         freeze_with_config(cfg, db=self.db)
         self.assertRaises(FreezeException, freeze_export, Export(cfg.data['common'], {'query': 'SELECT * FROM notable'}))
+
+    def test_unicode_path(self):
+        cfg = Configuration(os.path.join(os.path.dirname(__file__), 'Freezefile.yaml'))
+        cfg.data['common']['database'] = self.db
+        cfg.data['common']['prefix'] = os.path.join(self.d, u'über')
+        cfg.data['common']['query'] = 'SELECT * FROM weather'
+        cfg.data['exports'] = [{'filename': 'weather.csv', 'format': 'csv'}]
+        freeze_with_config(cfg, db=self.db)
+
+
+if __name__ == '__main__':
+    unittest.main()